Moved ParseTimepoint to its own header
