+++ format.sh
@@ -1,3 +1,3 @@
 #!/bin/bash
 
-clang-format -i `find src -name *.cpp` `find src -name *.h`
+clang-format -i `find src -name *.cpp` `find include -name *.h`
